Make query variables 16 bits, not 32
Every free variable turns into a loop, so you don't want to have many of
these anyway. If you make a query with 2^32 variables, you are going to
have a bad time. Even 2^16 is far too much. On the other hand, I can
imagine that in very large queries you might end up with 256 variables.
That would still be a bad idea, but it is a very low bound. If we are
going to have an arbitrary limit, let's make it 2^16.

The reason I'm doing this now is because it will make the serialization
format more compact.

+++ src/query.rs
@@ -21,7 +21,7 @@ use types::Type;
 /// Variables can be named, but the names are specified separately; internally
 /// variables are referenced by index.
 #[derive(Copy, Clone, Debug, Eq, Hash, Ord, PartialEq, PartialOrd)]
-pub struct Var(pub u32);
+pub struct Var(pub u16);
 
 /// An attribute in a query can be named, or a fixed attribute id.
 ///
